organized js files into three folders components, other, and routes
