+++ src/routes/Home.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "./App.css";
-import Navbar from './Navbar';
-import Footer from './Footer';
+import "../App.css";
+import Navbar from '../components/Navbar';
+import Footer from '../components/Footer';
 
 function Home() {
   return (
