+++ src/routes/Order.js
@@ -1,7 +1,7 @@
 import React from "react";
-import Footer from "./Footer";
-import "./App.css";
-import Navbar from "./Navbar";
+import Footer from "../components/Footer";
+import "../App.css";
+import Navbar from "../components/Navbar";
 
 function Order(){
     return (
